requirements.txt
# Premier AI Research Lab - Dependencies
# Energy-efficient ML training and inference research

# Core ML
torch>=2.0.0
numpy>=1.24.0

# Experiment tracking
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Energy & carbon tracking
codecarbon>=2.3.0
psutil>=5.9.0

# Model compression
bitsandbytes>=0.41.0

# Visualization
tensorboard>=2.13.0

# API client
requests>=2.28.0
python-dotenv>=1.0.0

# Transformers (optional, for advanced experiments)
# transformers>=4.30.0
# accelerate>=0.20.0
